New endpoint for getting accepted recipes of a user, updated recipe endpoints order, updated SQL file.
Added new endpoint for getting accepted recipes of a user by username.
Modified order of endpoints and functions in recipes to be more logical.
Updated SQL file to contain new procedure for the new endpoint, renamed some older procedures to be more explicit, and modified procedures order to be more logical.