Frontend pagination and smaller fixes
Implemented pagination for every recipes page, profile page and admin page.
Fixed a bug where if the user modified the username while on the route /profile/username the username in the route wouldn't change to the new username.
Fixed a bug with recipe names on recipe cards so long recipe names no longer go out of bound.
Added quality of life functionality so when the filter menu is open if the user click outside of it the filter menu is closed.